Fix cache-deletes-data-prematurely issue
This commit fixes a _very_ awkward to debug issue which could not be reproduced for many days in the development environment, but occurred all the time in production. The cached data was basically just gone for some reason after 5 minutes, and there was no explanation available nor logs that would've explained the behavior.

Flask-Caching thinks it's a good idea to hardcode a default timeout for the cached data. Our application requires that data is stored at least until the error timeout has triggered (1h) to allow one request handler to update it with new data.
For some reason, the author does not want to introduce a possibility to disable timeouts by default, or to set timeout to infinity. Therefore, we just set it to some absurdly high number (2^31 at this point, which equals ~68 years) to achieve basically the same goal.

Note for self: Python 3 no longer offers sys.maxint, and no, sys.maxsize is _not_ the same thing.